Moved object-utils file
